requirements.txt
playwright
beautifulsoup4
rich
click
lxml
psycopg2-binary